Move project tenant controller (#116)
* modify svc const (#112)

* move tenant and project controller to warden

* support sync cluster without scout

* fix pkg import
