+++ pkg/warden/localmgr/setup.go
@@ -22,8 +22,12 @@ import (
 
 	"github.com/kubecube-io/kubecube/pkg/warden/localmgr/controllers/hotplug"
 	"github.com/kubecube-io/kubecube/pkg/warden/localmgr/controllers/olm"
+	project "github.com/kubecube-io/kubecube/pkg/warden/localmgr/controllers/project"
+	tenant "github.com/kubecube-io/kubecube/pkg/warden/localmgr/controllers/tenant"
 	hotplug2 "github.com/kubecube-io/kubecube/pkg/warden/localmgr/webhooks/hotplug"
+	project2 "github.com/kubecube-io/kubecube/pkg/warden/localmgr/webhooks/project"
 	"github.com/kubecube-io/kubecube/pkg/warden/localmgr/webhooks/quota"
+	tenant2 "github.com/kubecube-io/kubecube/pkg/warden/localmgr/webhooks/tenant"
 )
 
 // setupControllersWithManager set up controllers into manager
@@ -40,6 +44,16 @@ func setupControllersWithManager(m *LocalManager) error {
 		return err
 	}
 
+	err = tenant.SetupWithManager(m.Manager)
+	if err != nil {
+		return err
+	}
+
+	err = project.SetupWithManager(m.Manager)
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
 
@@ -47,6 +61,8 @@ func setupControllersWithManager(m *LocalManager) error {
 func setupWithWebhooks(m *LocalManager) {
 	hookServer := m.GetWebhookServer()
 
+	hookServer.Register("/validate-tenant-kubecube-io-v1-tenant", admisson.ValidatingWebhookFor(tenant2.NewTenantValidator(m.GetClient())))
+	hookServer.Register("/validate-tenant-kubecube-io-v1-project", admisson.ValidatingWebhookFor(project2.NewProjectValidator(m.GetClient())))
 	hookServer.Register("/validate-core-kubernetes-v1-resource-quota", &webhook.Admission{Handler: &quota.ResourceQuotaValidator{PivotClient: m.PivotClient.Direct(), LocalClient: m.GetClient()}})
 	hookServer.Register("/warden-validate-hotplug-kubecube-io-v1-hotplug", admisson.ValidatingWebhookFor(hotplug2.NewHotplugValidator(m.IsMemberCluster)))
 }
